backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
pymysql
pyjwt
bcrypt
python-dotenv
werkzeug
flask-limiter
cryptography
